# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AGENT STATS",
    "description": "AGENT STATS STATUSES AND CHANNEL DATA"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/signIn": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to sign in a specific user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/user/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get a specific user.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User registered successfully.",
            "schema": {
              "$ref": "#/definitions/Authorized"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/job/": {
      "post": {
        "tags": [
          "Job"
        ],
        "description": "create new out dial jobs.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateJob"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "schema": {
              "$ref": "#/definitions/SipPeers"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/{id}": {
      "put": {
        "tags": [
          "Job"
        ],
        "description": "update outbound job by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateJob"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "schema": {
              "$ref": "#/definitions/SipPeers"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "schema": {
              "$ref": "#/definitions/SipPeers"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/job/types": {
      "get": {
        "tags": [
          "Job"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "schema": {
              "$ref": "#/definitions/SipPeers"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
...
    "CreateJob": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "test"
        },
        "type": {
          "type": "string",
          "example": "predictive"
        },
        "pause": {
          "type": "number",
          "example": 0
        },
        "complete": {
          "type": "number",
          "example": 0
        },
        "startTime": {
          "type": "string",
          "example": "2022-08-10 20:00:00"
        },
        "endTime": {
          "type": "string",
          "example": "2022-08-10 23:00:00"
        },
        "jobData": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "number": {
                "type": "string",
                "example": "231203"
              },
              "mask": {
                "type": "string",
                "example": "123456"
              },
              "context": {
                "type": "string",
                "example": "default"
              },
              "trunk": {
                "type": "string",
                "example": "outgoing"
              }
            }
          }
        }
      }
    },
    "UpdateJob": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "test"
        },
        "type": {
          "type": "string",
          "example": "predictive"
        },
        "pause": {
          "type": "number",
          "example": 0
        },
        "complete": {
          "type": "number",
          "example": 0
        },
        "startTime": {
          "type": "string",
          "example": "2022-08-10 20:00:00"
        },
        "endTime": {
          "type": "string",
          "example": "2022-08-10 23:00:00"
        },
        "jobData": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "number": {
                "type": "string",
                "example": "231203"
              },
              "mask": {
                "type": "string",
                "example": "123456"
              },
              "context": {
                "type": "string",
                "example": "default"
              },
              "trunk": {
                "type": "string",
                "example": "outgoing"
              }
            }
          }
        }
      }
    },
...
    "SipPeers": {
      "type": "object",
      "properties": {
        "response": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": true
            },
            "data": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "event": {
                    "type": "string",
                    "example": "PeerEntry"
                  },
                  "actionid": {
                    "type": "string",
                    "example": "1660286907216"
                  },
                  "channeltype": {
                    "type": "string",
                    "example": "SIP"
                  },
                  "objectname": {
                    "type": "string",
                    "example": "salman"
                  },
                  "chanobjecttype": {
                    "type": "string",
                    "example": "peer"
                  },
                  "ipaddress": {
                    "type": "string",
                    "example": "-none-"
                  },
                  "ipport": {
                    "type": "string",
                    "example": "0"
                  },
                  "dynamic": {
                    "type": "string",
                    "example": "yes"
                  },
                  "autoforcerport": {
                    "type": "string",
                    "example": "no"
                  },
                  "forcerport": {
                    "type": "string",
                    "example": "no"
                  },
                  "autocomedia": {
                    "type": "string",
                    "example": "no"
                  },
                  "comedia": {
                    "type": "string",
                    "example": "no"
                  },
                  "videosupport": {
                    "type": "string",
                    "example": "no"
                  },
                  "textsupport": {
                    "type": "string",
                    "example": "no"
                  },
                  "acl": {
                    "type": "string",
                    "example": "no"
                  },
                  "status": {
                    "type": "string",
                    "example": "UNKNOWN"
                  },
                  "realtimedevice": {
                    "type": "string",
                    "example": "no"
                  },
                  "description": {
                    "type": "string",
                    "example": ""
                  },
                  "accountcode": {
                    "type": "string",
                    "example": ""
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "AddUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "BEN@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "XYZ"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "Authorized": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Successfully found"
        }
      }
    }
  }
}